// web/src/types/SBNews.d.ts
import {
	AssetStoryblok,
	RelatedLinkStoryblok,
	RelatedNewsLinkStoryblok,
} from "./storyblok";

import { StoryblokRichtext } from "storyblok-rich-text-react-renderer";

/** NOTE: Because we are using the storyblok-rich-text-react-renderer,
 * we need to import the type from the package and use this in
 * place of the RichTextStoryblok type that is generated by storyblok.
 * Other types we need are still imported from the generated types from storyblok */

export interface SBNewsArticle {
	title: string;
	date: string;
	introText: string;
	content: StoryblokRichtext;
	image: AssetStoryblok;
	resources?: RelatedLinkStoryblok[];
	relatedNews?: RelatedNewsLinkStoryblok[];
	_uid: string;
	component: "newsArticle";
	[k: string]: any;
}

/** TODO: Handle the any type or allow it in this file? */